pyproject.toml
[project]
name = "adk"
version = "0.1.0"
description = "Google Agent Development Kit (ADK) Examples"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=0.1.0",
    "litellm>=1.66.2",
    "load-dotenv>=0.1.0",
    "logging>=0.4.9.6",
]
